fix: ptr-ptr conversions
These are causing dc1 to emit warnings, but warnings cause an
Amiga-style return code of 5. Unix Make considers that an error and
abort the build.

Could maybe configure make to not worry about a return code of 5, but
why not just fix the warnings? :-)

+++ src/dcpp/cpp.c
@@ -126,7 +126,7 @@ int32_t xbytes;
                 }
             }
 #else
-            len = HandleTriGraphs(base) - base;
+            len = HandleTriGraphs(base) - (int)base;
 #endif
         }
     } else {             /*  else macro  */
